Refactor la gestion de la connexion à la base de données et mise à jour des chemins d'inclusion
