+++ index.php
@@ -5,7 +5,7 @@ session_name("main");
 session_start();
 
 // Inclusion de fichier nécessaire 
-require_once 'src/control/BDDControl/connectBDD.php';
+require_once './src/controller/bddController/connectBDD.php';
 
 // Vérifie les paramètres après ? dans l'URL, si vide redirection vers la page d'accueil
 if (empty($_SERVER['QUERY_STRING'])) {
